document-summarization/lambda/requirements.txt
boto3==1.34.0
PyPDF2==3.0.1
python-docx==1.1.0
